+++ lib/edart_builder.dart
@@ -1,7 +1,5 @@
 import 'package:build/build.dart';
 import 'package:edart/edart_compiler.dart';
-import 'package:edart/edart_parser.dart';
-import 'package:edart/fragment.dart';
 import 'package:path/path.dart' as _path;
 
 Builder edartBuilder(BuilderOptions options) => EdartBuilder();
@@ -17,17 +15,12 @@ class EdartBuilder implements Builder {
     final inputId = buildStep.inputId;
     final extension = _path.extension(inputId.path);
     final outputId = buildStep.inputId.changeExtension('${extension}.g.dart');
-    final data = await buildStep.readAsString(inputId);
-    final parser = EdartParser();
-    final fragments = parser.parse(data) as List<Fragment>;
-    if (parser.error != null) {
-      throw parser.error;
-    }
-
+    final source = await buildStep.readAsString(inputId);
     final compiler = EdartCompiler();
     final path = inputId.path;
     final classname = _path.basename(path).replaceAll('.', '_');
-    final code = compiler.compile(path, classname, fragments);
+    final code =
+        compiler.compile(classname: classname, filename: path, source: source);
     await buildStep.writeAsString(outputId, code);
   }
 }