+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.1.1
 
 - Added build transformer `EdartBuilder` (`edart_builder.dart`)
+- Method `EdartCompiler.compile()` renamed to `EdartCompiler.compileFragments()`
+- Added new method `EdartCompiler.compile()`
 
 ## 0.1.0